+++ src/test/java/com/arangodb/example/graph/BaseGraphTest.java
@@ -23,10 +23,11 @@ package com.arangodb.example.graph;
 import com.arangodb.ArangoDB;
 import com.arangodb.ArangoDBException;
 import com.arangodb.ArangoDatabase;
+import com.arangodb.DbName;
 import com.arangodb.entity.EdgeDefinition;
 import com.arangodb.entity.VertexEntity;
-import org.junit.AfterClass;
-import org.junit.BeforeClass;
+import org.junit.jupiter.api.AfterAll;
+import org.junit.jupiter.api.BeforeAll;
 
 import java.util.ArrayList;
 import java.util.Collection;
@@ -34,17 +35,17 @@ import java.util.Collection;
 /**
  * @author Mark Vollmary
  */
-public abstract class BaseGraphTest {
+abstract class BaseGraphTest {
 
-    private static final String TEST_DB = "java_driver_graph_test_db";
+    private static final DbName TEST_DB = DbName.of("java_driver_graph_test_db");
     private static ArangoDB arangoDB;
     static ArangoDatabase db;
     private static final String GRAPH_NAME = "traversalGraph";
     private static final String EDGE_COLLECTION_NAME = "edges";
-    private static final String VERTEXT_COLLECTION_NAME = "circles";
+    private static final String VERTEX_COLLECTION_NAME = "circles";
 
-    @BeforeClass
-    public static void init() {
+    @BeforeAll
+    static void init() {
         if (arangoDB == null) {
             arangoDB = new ArangoDB.Builder().build();
         }
@@ -55,15 +56,15 @@ public abstract class BaseGraphTest {
 
         final Collection<EdgeDefinition> edgeDefinitions = new ArrayList<>();
         final EdgeDefinition edgeDefinition = new EdgeDefinition().collection(EDGE_COLLECTION_NAME)
-                .from(VERTEXT_COLLECTION_NAME).to(VERTEXT_COLLECTION_NAME);
+                .from(VERTEX_COLLECTION_NAME).to(VERTEX_COLLECTION_NAME);
         edgeDefinitions.add(edgeDefinition);
         if (!db.graph(GRAPH_NAME).exists())
             db.createGraph(GRAPH_NAME, edgeDefinitions, null);
         addExampleElements();
     }
 
-    @AfterClass
-    public static void shutdown() {
+    @AfterAll
+    static void shutdown() {
         arangoDB.db(TEST_DB).drop();
         arangoDB.shutdown();
         arangoDB = null;
@@ -104,7 +105,7 @@ public abstract class BaseGraphTest {
     }
 
     private static VertexEntity createVertex(final Circle vertex) throws ArangoDBException {
-        return db.graph(GRAPH_NAME).vertexCollection(VERTEXT_COLLECTION_NAME).insertVertex(vertex);
+        return db.graph(GRAPH_NAME).vertexCollection(VERTEX_COLLECTION_NAME).insertVertex(vertex);
     }
 
 }
